hypervisor: arm: phys_processor_id should return logical id
Currently the function phys_processor_id identifies the current
CPU by reading the MPIDR register. However, on systems with
multiple implemented affinity levels, the scheme used by this
register is hierarchical and does not correspond to the logical
IDs allocated to CPUs.

Change phys_processor_id to return the logical CPU id, so we
don't run into problems when we start implementing support for
affinity levels.

Signed-off-by: Antonios Motakis <antonios.motakis@huawei.com>